Fix multiple CommitSchedulers (#82)
The ContextVar is thread-local (one per thread), so using that as the
source of what should be a single global CommitScheduler was the source
of this bug. This commit moves the CommitScheduler to a class variable
protected by a thread lock, and there is now a single global
CommitScheduler.

This should fix any 429/500 errors from commits.