Only persist to Dataset if running in a Space
We can fix both #78 and #80 if we limit Dataset persistence to only
those trackio instances running in Spaces. This is both simpler from a
user perspective (fewer combinations of different behavior for
space/dataset/project) and simpler to implement vs. supporting multiple
projects w/ potentially different Datasets to persist to.